[ADD] Day 1 (Final Sum)
Open a File ...txt
Iterate through the file's Lines
Taking the First and Last digit of the line and adding it to the Global Summary

Global Summary + CurrentLine
Print the Global Summary at that time.

Continue iterating to the end of the Lines and print the FInal Global Summary.

Comments written by " Github co-Pilot "